+++ src/test/java/tudelft/domain/TwoIntegerTest.java
@@ -1,8 +1,9 @@
-package tudelft.twointeger1;
+package tudelft.domain;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import tudelft.domain.TwoInteger;
 
 class TwoIntegerTest {
